chore(weave): extract clickhouse/costs.py with CHInfraProtocol
Move cost CRUD (create, query, purge) and cost computation helpers
into CostsMixin.
